Fix bug - mario_raccoon's tail destroys everything in sight
Cause: The tail's size is a little bit too big and also in the past, mario was implemented to be briefly invincible when he whips tails, which is not really the case for the original game

+++ SuperMarioBros3/AttackBBox.cpp
@@ -15,7 +15,7 @@
 #include "PBlock.h"
 #include "Coin.h"
 
-CAttackBBox::CAttackBBox(float x, float y, float vx, float vy, int width, int height, 
+CAttackBBox::CAttackBBox(float x, float y, float vx, float vy, float width, float height, 
 	const LPPLAYSCENE currentScene) : CGameObject(x, y), currentScene(currentScene)
 {
 	this->vx = vx;
